cli: split commands/config.rs into sub modules
