+++ cli/src/commands/config/get.rs
@@ -1,0 +1,76 @@
+// Copyright 2020 The Jujutsu Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::io::Write as _;
+
+use tracing::instrument;
+
+use crate::cli_util::CommandHelper;
+use crate::command_error::{config_error, CommandError};
+use crate::config::ConfigNamePathBuf;
+use crate::ui::Ui;
+
+/// Get the value of a given config option.
+///
+/// Unlike `jj config list`, the result of `jj config get` is printed without
+/// extra formatting and therefore is usable in scripting. For example:
+///
+/// $ jj config list user.name
+/// user.name="Martin von Zweigbergk"
+/// $ jj config get user.name
+/// Martin von Zweigbergk
+#[derive(clap::Args, Clone, Debug)]
+#[command(verbatim_doc_comment)]
+pub struct ConfigGetArgs {
+    #[arg(required = true)]
+    name: ConfigNamePathBuf,
+}
+
+#[instrument(skip_all)]
+pub fn cmd_config_get(
+    ui: &mut Ui,
+    command: &CommandHelper,
+    args: &ConfigGetArgs,
+) -> Result<(), CommandError> {
+    let value = args
+        .name
+        .lookup_value(command.settings().config())
+        .and_then(|value| value.into_string())
+        .map_err(|err| match err {
+            config::ConfigError::Type {
+                origin,
+                unexpected,
+                expected,
+                key,
+            } => {
+                let expected = format!("a value convertible to {expected}");
+                // Copied from `impl fmt::Display for ConfigError`. We can't use
+                // the `Display` impl directly because `expected` is required to
+                // be a `'static str`.
+                let mut buf = String::new();
+                use std::fmt::Write;
+                write!(buf, "invalid type: {unexpected}, expected {expected}").unwrap();
+                if let Some(key) = key {
+                    write!(buf, " for key `{key}`").unwrap();
+                }
+                if let Some(origin) = origin {
+                    write!(buf, " in {origin}").unwrap();
+                }
+                config_error(buf)
+            }
+            err => err.into(),
+        })?;
+    writeln!(ui.stdout(), "{value}")?;
+    Ok(())
+}
